Implemented functionality to check current ingredient inventory when browsing public recipes, to check if the user has the correct ingredients to make the recipe.

+++ app/src/main/java/com/tm470/cookhub/models/Quantity.kt
@@ -1,6 +1,6 @@
 package com.tm470.cookhub.models
 
-class Quantity(var amount: Int, var type: String) {
+class Quantity(var amount: Double, var type: String) {
 
-    constructor(): this(0, "g")
+    constructor(): this(0.0, "g")
 }
